# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Webhook Laravel",
        "description": "API pour la gestion et l'envoi de webhooks",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/webhooks/trigger/{eventType}": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Déclencher un webhook",
                "operationId": "58247496226ec66269aee1920883a1f7",
                "parameters": [
                    {
                        "name": "eventType",
                        "in": "path",
                        "description": "Type d'événement à déclencher",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "user.created"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook déclenché avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook dispatched"
                                        },
                                        "event_type": {
                                            "type": "string",
                                            "example": "user.created"
                                        },
                                        "payload": {
                                            "type": "object",
                                            "example": {
                                                "event": "user.created",
                                                "timestamp": "2023-12-07T10:30:00.000000Z",
                                                "data": {
                                                    "message": "Webhook triggered successfully",
                                                    "id": "655f7a8b1a2b3"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Type d'événement invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid event type"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun webhook trouvé pour cet événement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No webhooks found for this event type"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}